Move specific log packages to their modules
