+++ python_utils/log/formatters.py
@@ -102,47 +102,3 @@ class BasicRequestGELFFormatter(BasicGELFFormatter):
             }
         }
         super().__init__(extra_fields=extra_fields)
-
-
-class DjangoRequestGELFFormatter(BasicRequestGELFFormatter):
-    """A GELF formatter to format a :class:`logging.LogRecord` into GELF, with specific request fields"""
-
-    def set_more_extra_fields(self, record):
-        """Transform from :class:`logging.LogRecord` some fields to python dict.
-
-        :param :class:`logging.LogRecord` record: record emitted by logger. Its attribures can be seen
-        in https://docs.python.org/3/library/logging.html#logrecord-attributes
-        """
-        request = getattr(record, 'request', None)
-        if request:
-            setattr(record, 'scheme', getattr(record.request, 'scheme', None))
-            self.extra_fields['scheme'] = True
-
-            setattr(record, 'method', getattr(record.request, 'method', None))
-            self.extra_fields['method'] = True
-
-            user = getattr(record.request, 'user', None)
-            if user:
-                setattr(record, 'user_id', getattr(user, 'id', None))
-                self.extra_fields['user_id'] = True
-
-                setattr(record, 'username', getattr(user, 'username', None))
-                self.extra_fields['username'] = True
-
-
-class GunicornRequestGELFFormatter(BasicGELFFormatter):
-    """A GELF formatter to format a :class:`logging.LogRecord` into GELF, adding gunicorn access log specific fields"""
-
-    def __init__(self):
-        extra_fields = {
-            **EXTRA_FIELDS,
-            **{
-                'remote_addr': True,
-                'response_status_code': True,
-                'http_user_agent': True,
-                'request_method': True,
-                'path_info': True,
-                'server_protocol': True
-            }
-        }
-        super().__init__(extra_fields=extra_fields)
